--- pycaret/utils/patches.py
@@ -1,39 +1,0 @@
-from typing import Callable
-from unittest.mock import patch
-
-from sklearn.model_selection._validation import _fit_and_score, _score
-
-
-def fit_and_score(*args, **kwargs) -> dict:
-    """Wrapper for sklearn's _fit_and_score function.
-
-    Wrap the function sklearn.model_selection._validation._fit_and_score
-    to, in turn, path sklearn's _score function to accept pipelines that
-    drop samples during transforming, within a joblib parallel context.
-
-    """
-
-    def wrapper(*args, **kwargs) -> dict:
-        with patch("sklearn.model_selection._validation._score", score(_score)):
-            return _fit_and_score(*args, **kwargs)
-
-    return wrapper(*args, **kwargs)
-
-
-def score(f: Callable) -> Callable:
-    """Patch decorator for sklearn's _score function.
-
-    Monkey patch for sklearn.model_selection._validation._score
-    function to score pipelines that drop samples during transforming.
-
-    """
-
-    def wrapper(*args, **kwargs):
-        args = list(args)  # Convert to list for item assignment
-        if len(args[0]) > 1:  # Has transformers
-            args[1], args[2] = args[0][:-1].transform(args[1], args[2])
-
-        # Return f(final_estimator, X_transformed, y_transformed, ...)
-        return f(args[0][-1], *tuple(args[1:]), **kwargs)
-
-    return wrapper
